Refactor authentication to be route-based and split home page
This commit introduces a new route-based authentication system and refactors the home page to separate public content from the protected editor.

Key changes:
- A new `src/config/routes.ts` file now defines public routes, centralizing the authentication logic.
- `AuthGuard.tsx` has been updated to use this new configuration, protecting routes that are not marked as public.
- The `useUser` hook now has a loading state to prevent UI flashes during authentication checks.
- The `Home.tsx` component has been split into a public-facing `Home` component and a new protected `Editor` component.
- A new `/editor` route has been created for the image editor.
- Image data is passed from the home page to the editor via `sessionStorage`.
- A "Sign Up" link has been added to the login form for better user navigation.
