+++ src/app/editor/page.tsx
@@ -1,0 +1,10 @@
+import { Editor } from "@/components/Editor";
+import AuthGuard from "@/components/AuthGuard";
+
+export default function EditorPage() {
+  return (
+    <AuthGuard>
+      <Editor />
+    </AuthGuard>
+  );
+}
