+++ src/app/page.tsx
@@ -1,10 +1,5 @@
 import { Home as HomePage } from "@/components/Home";
-import AuthGuard from "@/components/AuthGuard";
 
 export default function Home() {
-  return (
-    <AuthGuard>
-      <HomePage />
-    </AuthGuard>
-  );
+  return <HomePage />;
 }
