+++ src/config/routes.ts
@@ -1,0 +1,5 @@
+export const publicRoutes: string[] = ["/", "/login", "/signup"];
+
+export const isProtectedRoute = (path: string): boolean => {
+  return !publicRoutes.includes(path);
+};
